examples/simple-cloudevent/flask-pub-ce-demo/requirements.txt
Flask
python-dotenv

git+https://github.com/daxiom/simple-cloudevent.py.git@0.0.1
git+https://github.com/daxiom/flask-pub.git@0.0.2
